requirements.txt
flask==1.1.1
google-cloud-firestore==1.6.2